% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigwoods}
\alias{bigwoods}
\title{Bigwoods forest data}
\format{A data frame with 50023 rows and 10 variables:
\describe{
\item{ID}{Tree ID}
\item{species}{Linnaean species name}
\item{family_phylo}{Corresponding species phylogenetic family}
\item{trait_group}{Corresponding species trait-based cluster}
\item{x}{x-coordinate (meters from reference point)}
\item{y}{y-coordinate (meters from reference point)}
\item{growth}{Average annual growth in diameter at breast height from 2008-2014 (6 years)}
\item{dbh08}{Diameter at breast high in 2008}
\item{dhb14}{Diameter at breast high in 2014}
\item{code14}{Code}
}}
\usage{
bigwoods
}
\description{
Census of trees in bigwoods forest
}
\seealso{
\code{\link{families}}
}
\keyword{datasets}
