% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigwoods_study_region}
\alias{bigwoods_study_region}
\title{Bigwoods forest study region boundary}
\format{A data frame with 13 rows and 2 variables:
\describe{
\item{x}{x-coordinate (meters from reference point)}
\item{y}{y-coordinate (meters from reference point)}
}}
\usage{
bigwoods_study_region
}
\description{
Boundary region for bigwoods defined in terms of (x,y) vertices of a polygon.
}
\examples{
{
library(ggplot2)
ggplot(data = bigwoods, aes(x,y)) +
  geom_point() +
  coord_fixed(ratio=1) +
  geom_path(data = study_region_vertices, col = "red", size = 1)
}
}
\seealso{
\code{\link{families}}
}
\keyword{datasets}
