% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bw_census_2014}
\alias{bw_census_2014}
\title{Michigan Big Woods research plot data}
\format{
A data frame with 48371 rows and 8 variables:
\describe{
\item{treeID}{Tree identification number. This identifies an individual tree and
can be used to connect trees between the two censuses.}
\item{stemID}{Stem number for a multi-stemmed individual. For all trees this starts
at 1 and continues up from there. To uniquely identify a stem across the plot this
value must be combined with \code{treeID}.}
\item{sp}{Code for the species. See \code{\link{bw_species}} for scientific name.}
\item{gx}{x-coordinate meters from reference point}
\item{gy}{y-coordinate meters from reference point}
\item{date}{Approximate date the stem was measured}
\item{codes}{Code for additional informaiton on the stem: M means the main stem
of the individual tree; AL means the stem is alive but leaning or completely fallen
over; B means the stem is broken and over half the canopy is assumed to be missing;
and R means the stem was lost, but the tag was moved to another stem greater than DBH cutoff,
this stands for resprout.}
}
}
\usage{
bw_census_2014
}
\description{
The \href{https://deepblue.lib.umich.edu/data/concern/data_sets/ht24wj48w}{Big Woods}
data come from three censuses of a 23 ha forest research plots. All
free-standing vegetation greater than 1 cm diameter at 1.3 m height (diameter
at breast height; DBH) were tagged, identified, spatially mapped and had
their DBH measured. The original census took place in 2003 and covered only
12 ha. A second census took place from 2008-2010 and expanded the plot to its
current 23 ha. In the first and second censuses trees larger than 3.1 cm DBH
were included. Finally a third census took place in 2014. In this census
trees larger than 1 cm DBH were included. In the second and third censuses
the original trees were found, recorded for survival, remeasured, and new
individuals were tagged. This data frame has data from the third census
(2014).
}
\details{
This plot is part of the Smithsonian Institution's Forest Global Earth
Observatory \href{https://forestgeo.si.edu/}{(ForestGEO)} global network of
forest research sites. For complete details on this dataset see its
\href{https://deepblue.lib.umich.edu/data/concern/data_sets/ht24wj48w}{Deep Blue Data repository page}.
}
\examples{
data("bw_census_2008","bw_census_2014")
library(tidyverse)
# species-specific mortality between censuses
bw_census_2008 \%>\%
 left_join(bw_census_2014, by = c('treeID', 'stemID'), suffix = c("_1", "_2") ) \%>\%
 mutate(mortality = ifelse(is.na(dbh_2),1,0)) \%>\%
 group_by(sp_1) \%>\%
 summarize(mortality = mean(mortality), n = n()) \%>\%
 arrange(desc(n))
}
\seealso{
\code{\link{bw_species}} \code{\link{bw_census_2014}} \code{\link{compute_growth}}
}
\keyword{datasets}
