% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bw_censusdf2}
\alias{bw_censusdf2}
\title{Michigan Big Woods research plot data}
\format{A data frame with 48371 rows and 8 variables:
\describe{
\item{treeID}{Tree identification number}
\item{stemID}{Stem number for a multi-stemmed individual}
\item{sp}{Code for the species. See \code{bw_species} for scientific name.}
\item{gx}{x-coordinate meters from reference point}
\item{gy}{y-coordinate meters from reference point}
\item{date}{Approximate date the stem was measured}
\item{codes}{Code for additional informaiton on the stem}
}}
\usage{
bw_censusdf2
}
\description{
The \href{https://deepblue.lib.umich.edu/data/concern/data_sets/ht24wj48w}{Big Woods}
data come from three censuses of a 23 ha forest research plots. All
free-standing vegetation greater than 1 cm diameter at 1.3 m height (diameter
at breast height; DBH) were tagged, identified, spatially mapped and had
their DBH measured. The original census took place in 2003 and covered only
12 ha. A second census took place from 2008-2010 and expanded the plot to its
current 23 ha. In the first and second censuses trees larger than 3.1 cm DBH
were included. Finally a third census took place in 2014. In this census trees
larger than 1 cm DBH were included. In the second and third censuses the original
trees were found, recorded for survival, remeasured, and new individuals were tagged.
This data frame has data from the third census (2014).
}
\details{
This plot is part of the Smithsonian Institution's Forest Global Earth
Observatory \href{https://forestgeo.si.edu/}{(ForestGEO)} global network of
forest research sites.
}
\seealso{
\code{\link{bw_species}} \code{\link{bw_censusdf2}} \code{\link{compute_growth}}
}
\keyword{datasets}
