% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bw_study_region}
\alias{bw_study_region}
\title{Bigwoods forest study region boundary}
\format{
A \code{sf} spatial polygon.
\describe{
\item{x}{x-coordinate (meters from reference point)}
\item{y}{y-coordinate (meters from reference point)}
}
}
\usage{
bw_study_region
}
\description{
Boundary region for bigwoods defined in terms of (x,y) vertices of a polygon.
}
\examples{
library(ggplot2)
ggplot(bw_census_2008, aes(x = gx, y = gy)) +
  # Mark study region boundary
  geom_path(data = bw_study_region, size = 1) +
  coord_fixed(ratio = 1)
data("bw_census_2008","bw_study_region")
library(tidyverse)
library(sf)
# plot stems with plot boundary
bw_census_2008 \%>\%
 st_as_sf(coords = c('gx', 'gy')) \%>\%
 ggplot() +
 geom_sf(size = 0.25) +
 geom_sf(data = bw_study_region, color = 'red', fill = 'transparent')
}
\seealso{
\code{\link{bw_census_2008}} \code{\link{define_buffer}}
}
\keyword{datasets}
