% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{census_df1_ex}
\alias{census_df1_ex}
\title{Example input census data for package use}
\format{
A \code{tibble}
\describe{
\item{ID}{Tree identification number. This identifies an individual tree and
can be used to connect trees between the two censuses.}
\item{sp}{Speices of the individual}
\item{gx}{x-coordinate meters from reference point}
\item{gy}{y-coordinate meters from reference point}
\item{date}{Date the stem was measured}
\item{codes}{Code for additional information on the stem: M means the main stem
of the individual tree and R means the stem was lost, but the tag was moved to another stem greater than DBH cutoff,
this stands for resprout.}
\item{dbh}{Diameter at breast hieght of the tree in cm}
}
}
\usage{
census_df1_ex
}
\description{
This is example forest census data to be analyzed with this package.
}
\examples{
data(census_df1_ex,census_df2_ex)
library(dplyr)
library(stringr)
library(snakecase)
# Filter out resprouts
census_df2_ex_no_r <- census_df2_ex \%>\%
 filter(!str_detect(codes, 'R'))
id <- 'ID'
ex_growth_df <-
 # Merge both censuses and compute growth:
 compute_growth(census_df1_ex, census_df2_ex_no_r, id) \%>\%
 mutate(
   sp = to_any_case(sp),
   sp = as.factor(sp))
}
\seealso{
\code{\link{census_df2_ex}}
}
\keyword{datasets}
