% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{comp_bayes_lm_ex}
\alias{comp_bayes_lm_ex}
\title{Example fit model}
\format{
??
\describe{
\item{??}{Something}
}
}
\usage{
comp_bayes_lm_ex
}
\description{
This has posterior parameters for the fit growth model.
}
\examples{
library(dplyr)
library(yardstick)

# Compare model predictions to observation
predictions <- comp_bayes_lm_ex \%>\%
  predict(focal_vs_comp = focal_vs_comp_ex) \%>\%
  right_join(ex_growth_df, by = c("focal_ID" = "ID"))

predictions \%>\%
  rmse(truth = growth, estimate = growth_hat) \%>\%
  pull(.estimate)

# Plot posterior parameters
comp_bayes_lm_ex \%>\%
  autoplot()
}
\seealso{
\code{\link{focal_vs_comp_ex}}
}
\keyword{datasets}
