% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{compute_buffer_region}
\alias{compute_buffer_region}
\title{Computer buffer to a region.}
\usage{
compute_buffer_region(region, direction = "in", size)
}
\arguments{
\item{region}{region to be buffered}

\item{direction}{"in" for buffers that are contained within \code{region}, "out" for buffers that contain \code{region}.}

\item{max_dist}{size of buffer in same units as geometry in \code{region}}
}
\value{
As \code{sf} object corresponding to buffer
}
\description{
Computer buffer to a region.
}
\examples{
library(tibble)
library(sfheaders)
library(ggplot2)

# Example square region to be buffered (as sf object)
region <- tibble(
  x = c(0, 0, 1, 1),
  y = c(0, 1, 1, 0)
) \%>\%
  sf_polygon()

# Size of buffer
size <- 0.05

# Compute "inwards" buffer
inwards_buffer_region <- region \%>\%
  compute_buffer_region(direction = "in", size = size)

# Compute "outwards" buffer
outwards_buffer_region <- region \%>\%
  compute_buffer_region(direction = "out", size = size)

# Plot original region in black, outwards buffer in blue, inwards buffer in orange
ggplot() +
  geom_sf(data = outwards_buffer_region, col = "blue") +
  geom_sf(data = region) +
  geom_sf(data = inwards_buffer_region, col = "orange")
}
