% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_focal_vs_comp}
\alias{create_focal_vs_comp}
\title{Create main data frame for analysis}
\usage{
create_focal_vs_comp(forest, max_dist, folds, notion_of_focal_species,
  notion_of_comp_species)
}
\arguments{
\item{forest}{data frame containing tree data}

\item{max_dist}{Maximum distance from a focal tree for another tree to be
considered a competitor tree}

\item{folds}{CV fold information from \code{\link{define_cv_grid}}}

\item{notion_of_focal_species}{from \code{\link{get_model_specs}}}

\item{notion_of_comp_species}{from \code{\link{get_model_specs}}}
}
\value{
\code{focal_vs_comp} data frame
}
\description{
Create main data frame for analysis
}
\examples{
1+1
}
\seealso{
\code{\link{define_cv_grid}} and \code{\link{get_model_specs}}
}
