% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_growth_df}
\alias{create_growth_df}
\title{Create growth_df from two census_df}
\usage{
create_growth_df(census_df1, census_df2, id, species_df)
}
\arguments{
\item{census_df1}{This is a table with data from the first census}

\item{census_df2}{This is a table with data from the second census}

\item{id}{The name of the varialbe in the census files which links individuals}

\item{species_df}{A table with data about each species}
}
\value{
growth_df data frame
}
\description{
Create growth_df from two census_df
}
\examples{
1+1
}
\seealso{

}
