% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{cv_grid_sf_ex}
\alias{cv_grid_sf_ex}
\title{Example cross validation grid}
\format{
A \code{sf} polygons
\describe{
\item{foldID}{Tree identification number. This identifies an individual tree and
can be used to connect trees between the two censuses.}
\item{geometry}{Point location of the individual}
}
}
\usage{
cv_grid_sf_ex
}
\description{
This is an example cross validation grid. This is needed to create the focal versus
comp data frame and run cross-validated models
}
\examples{
library(ggplot2)
library(sf)

max_dist <- 1

ggplot(cv_grid_sf_ex) +
  geom_sf() +
  geom_sf(data = growth_spatial_ex)

focal_vs_comp_ex <- growth_spatial_ex \%>\%
  create_focal_vs_comp(max_dist, cv_grid_sf = cv_grid_sf_ex, id = "ID")
}
\seealso{
\code{\link{growth_spatial_ex}}
}
\keyword{datasets}
