% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{define_bigwoods_buffer}
\alias{define_bigwoods_buffer}
\title{Identify which trees are in buffer region of bigwoods.}
\usage{
define_bigwoods_buffer(bigwoods, max_dist)
}
\arguments{
\item{bigwoods}{\code{\link{bigwoods}} data frame}

\item{max_dist}{Maximum distance from a focal tree for another tree to be
considered a competitor tree}
}
\value{
The same \code{\link{bigwoods}} data frame but with a new boolean
variable \code{buffer} indicating if a tree is in the study region buffer
area.
}
\description{
Identify which trees are in buffer region of bigwoods.
}
\examples{
1+1
}
\seealso{
\code{\link{bigwoods}}
}
