% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{define_cv_grid}
\alias{define_cv_grid}
\title{Define spatial crossvalidation grid}
\usage{
define_cv_grid(forest, cv_fold_size)
}
\arguments{
\item{forest}{data frame containing tree data}

\item{cv_fold_size}{grid size for spatial crossvalidation fold}
}
\value{
The same \code{forest} data frame but with a new numerical variable
\code{fold} indicating the fold number.
}
\description{
Define spatial crossvalidation grid
}
\examples{
library(ggplot2)
max_dist <- 7.5
cv_fold_size <- 100
bigwoods <- bigwoods \%>\%
  define_bigwoods_buffer(max_dist) \%>\%
  define_cv_grid(cv_fold_size)

ggplot(data = bigwoods, aes(x = x, y = y)) +
  geom_point(aes(col = factor(fold))) +
  coord_fixed(ratio = 1)
}
