% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ex_cv_grid_sf}
\alias{ex_cv_grid_sf}
\title{Example cross validation grid}
\format{
A \code{sf} polygons
\describe{
\item{foldID}{Tree identification number. This identifies an individual tree and
can be used to connect trees between the two censuses.}
\item{geometry}{Point location of the individual}
}
}
\usage{
ex_cv_grid_sf
}
\description{
This is an example cross validation grid. This is needed to create the focal versus
comp data frame and run cross-validated models
}
\seealso{
\code{\link{ex_growth_df_spatial}}
@examples
library(ggplot2)
library(sf)

ggplot(ex_cv_grid_sf) +
geom_sf() +
geom_sf(data = ex_growth_df_spatial)

focal_vs_comp_ex <- ex_growth_df_spatial \%>\%
create_focal_vs_comp(max_dist, cv_grid_sf = ex_cv_grid_sf, id = "ID")
}
\keyword{datasets}
