% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{ex_growth_df}
\alias{ex_growth_df}
\title{Example growth data frame for small example}
\format{
A \code{sf} spatial tibble
\describe{
\item{ID}{Tree identification number. This identifies an individual tree and
can be used to connect trees between the two censuses.}
\item{sp}{Species of the individual}
\item{codes1}{Code for additional information on the stem during the first census: M means the main stem
of the individual tree and R means the stem was lost, but the tag was moved to another stem greater than DBH cutoff,
this stands for resprout.}
\item{dbh1}{Diameter at breast height of the tree in cm at the first census}
\item{dbh2}{Diameter at breast height of the tree in cm at the second census}
\item{growth}{Average annual growth between the two censuses in cm per year}
\item{codes2}{Codes at the second census}
\item{geometry}{Point location of the individual}
}
}
\usage{
ex_growth_df
}
\description{
This is an example growth data frame formed from two census data frames. In this case it is
made by combining \code{\link{census_df1_ex}} and \code{\link{census_df2_ex}}. The
individuals alive in both censuses were linked by their tree ID.
}
\examples{
library(ggplot2)
library(dplyr)
library(sf)
library(sfheaders)
library(blockCV)

ex_growth_df \%>\%
  ggplot() +
  geom_sf()

ex_growth_df \%>\%
  group_by(sp) \%>\%
  summarize(mean(growth))

# Add buffer
ex_growth_df_spatial <- ex_growth_df \%>\%
  add_buffer_variable(direction = "in", size = 1, region = ex_study_region)

# Add cross-validation folds
fold1 <- rbind(c(0, 0), c(5, 0), c(5, 5), c(0, 5), c(0, 0))
fold2 <- rbind(c(5, 0), c(10, 0), c(10, 5), c(5, 5), c(5, 0))
blocks <- bind_rows(
  sf_polygon(fold1),
  sf_polygon(fold2)
) \%>\%
  mutate(foldID = c(1, 2))

ex_cv_grid <- spatialBlock(
  speciesData = ex_growth_df,
  verbose = FALSE,
  k = 2,
  selection = "systematic",
  blocks = blocks
)

# Add foldID to data
ex_growth_df_spatial <- ex_growth_df_spatial \%>\%
  mutate(foldID = ex_cv_grid$foldID \%>\% as.factor())
}
\seealso{
\code{\link{census_df1_ex}}, \code{\link{census_df2_ex}}, and \code{\link{compute_growth}}
}
\keyword{datasets}
