% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{ex_growth_df_spatial}
\alias{ex_growth_df_spatial}
\title{Example growth data frame with spatial data for small example}
\format{
A \code{sf} spatial tibble
\describe{
\item{ID}{Tree identification number. This identifies an individual tree and
can be used to connect trees between the two censuses.}
\item{sp}{Speices of the individual}
\item{codes1}{Code for additional information on the stem during the first census: M means the main stem
of the individual tree and R means the stem was lost, but the tag was moved to another stem greater than DBH cutoff,
this stands for resprout.}
\item{dbh1}{Diameter at breast hieght of the tree in cm at the first census}
\item{dbh2}{Diameter at breast hieght of the tree in cm at the second census}
\item{growth}{Average annual growth between the two censuses in cm per year}
\item{codes2}{Codes at the second census}
\item{geometry}{Point location of the individual}
\item{buffer}{A boolean variable for whether the individual is in the buffer region or not}
\item{foldID}{Which cross-validation fold the individual is in}
}
}
\usage{
ex_growth_df_spatial
}
\description{
This is an example growth data frame formed from two census data frames which
has been updated with spatial data. It starts from \code{\link{ex_growth_df}}.
}
\examples{
library(ggplot2)
library(dplyr)
library(sf)

max_dist <- 1

ggplot() +
  geom_sf(data = ex_growth_df_spatial, aes(col = buffer), size = 2)

ggplot() +
  geom_sf(data = ex_growth_df_spatial, aes(col = foldID), size = 2)

# Create the focal versus comp data frame
focal_vs_comp_ex <- ex_growth_df_spatial \%>\%
  create_focal_vs_comp(max_dist, cv_grid_sf = ex_cv_grid_sf, id = "ID")
}
\seealso{
\code{\link{ex_growth_df}}, \code{\link{add_buffer_variable}}, \code{\link{ex_cv_grid_sf}}
}
\keyword{datasets}
