% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{ex_study_region}
\alias{ex_study_region}
\title{Study region for example data}
\format{
A \code{sf} spatial features polygon
}
\usage{
ex_study_region
}
\description{
Boundary region for small example data set defined in terms of (x,y) vertices of a polygon.
}
\examples{
library(ggplot2)
library(sf)

# Convert stems to sf object
census_df1_ex_sf <- census_df1_ex \%>\%
  st_as_sf(coords = c("gx", "gy"))

# Plot stems with plot boundary
ggplot() +
  geom_sf(data = ex_study_region) +
  geom_sf(data = bw_study_region, color = "red", fill = "transparent")
}
\seealso{
\code{\link{census_df1_ex}}
}
\keyword{datasets}
