% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{families}
\alias{families}
\title{Phylogenic groupings and trait based clustering of various tree species}
\format{A data frame with 50 rows and 3 variables:
\describe{
\item{spcode}{Linnaean species name}
\item{family_phylo}{Taxonomic/phylogenetic grouping at higher level than family.}
\item{trait_group}{Clustering of species based on their traits rather than their evolutionary relationships.}
}}
\source{
See \url{https://en.wikipedia.org/wiki/Family_(biology)} for biological/taxonomical definition of family.
}
\usage{
families
}
\description{
A dataset mapping various tree species to their phylogenetic families and
trait-based (rather than evolutionary relationship-based) clusters.
}
\seealso{
\code{\link{bigwoods}}
}
\keyword{datasets}
