% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_bayesian_model}
\alias{fit_bayesian_model}
\title{Fit bayesian model}
\usage{
fit_bayesian_model(
  focal_vs_comp,
  model_specs,
  run_shuffle = FALSE,
  prior_hyperparameters = NULL
)
}
\arguments{
\item{focal_vs_comp}{from \code{\link{create_focal_vs_comp}}}

\item{model_specs}{from \code{\link{get_model_specs}}}

\item{run_shuffle}{boolean as to whether to run permutation test shuffle}

\item{prior_hyperparameters}{list of a0, b0, mu_0 and V_0}
}
\value{
Posterior parameter values
}
\description{
Fit bayesian model
}
\examples{
1+1
}
