% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestecology.R
\docType{package}
\name{forestecology}
\alias{forestecology}
\alias{forestecology-package}
\title{\code{forestecology} package}
\description{
Methods and data for forest ecology model selection and assessment
}
\details{
See the README on
\href{https://github.com/rudeboybert/forestecology#readme}{GitHub}
}
