% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_cv_fold_info}
\alias{get_cv_fold_info}
\title{Return information on each fold in crossvalidation grid}
\usage{
get_cv_fold_info(forest, cv_fold_size)
}
\arguments{
\item{forest}{data frame containing tree data}

\item{cv_fold_size}{grid size for spatial crossvalidation fold}
}
\value{
A "nested" dataframe with the fold number, its rough centroid, and
all its neighbors in a nested list.
}
\description{
Return information on each fold in crossvalidation grid
}
\examples{
1+1
}
