% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_model_specs}
\alias{get_model_specs}
\title{Define notion of species based on model choice}
\usage{
get_model_specs(forest, model_number)
}
\arguments{
\item{forest}{data frame containing tree data}

\item{model_number}{Number of model out of 1-5 from paper}
}
\value{
A list ouf outputs
}
\description{
Define notion of species based on model choice
}
\examples{
1+1
}
