% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{growth_df_ex}
\alias{growth_df_ex}
\title{Example input data for \code{\link{create_focal_vs_comp}}}
\format{
A \code{sf} spatial features polygon
}
\usage{
growth_df_ex
}
\description{
An example \code{sf} of type generated by \code{\link{compute_growth}}
}
\examples{
1 + 1
}
\keyword{datasets}
