% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{plot_bayesian_model_parameters}
\alias{plot_bayesian_model_parameters}
\title{Plot Bayesian model parameters}
\usage{
plot_bayesian_model_parameters(posterior_param, sp_to_plot = NULL)
}
\arguments{
\item{posterior_param}{Output of \code{\link{fit.fe_data}}}

\item{sp_to_plot}{Vector of subset of species to plot}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{.pred}
}
\description{
Plot Bayesian model parameters
}
\examples{
library(ggplot2)
library(ggridges)
# Load in posterior parameter example
data(posterior_param_ex)
plots <- plot_bayesian_model_parameters(posterior_param_ex)

# Plot beta_0, growth intercepts
plots[[1]]

# Plot beta_dbh, growth-dbh slopt
plots[[2]]

# Plot lambdas, competition coefficents
plots[[3]]
}
