% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{plot_beta0}
\alias{plot_beta0}
\title{Plot beta_0 parameters}
\usage{
plot_beta0(posterior_param, model_specs)
}
\arguments{
\item{posterior_param}{Output of \code{\link{fit_bayesian_model}}}

\item{model_specs}{from \code{\link{get_model_specs}}}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{growth_hat}
}
\description{
Plot beta_0 parameters
}
\examples{
1+1
}
