% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{plot_posterior_parameters}
\alias{plot_posterior_parameters}
\title{Plot beta_0 parameters}
\usage{
plot_posterior_parameters(posterior_param, sp_to_plot = NULL)
}
\arguments{
\item{posterior_param}{Output of \code{\link{fit_bayesian_model}}}

\item{sp_to_plot}{Vector of subset of species to plot}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{growth_hat}
}
\description{
Plot beta_0 parameters
}
\examples{
library(ggplot2)
library(ggridges)
# Load in posterior parameter example
data(posterior_param_ex)
plots <- plot_posterior_parameters(posterior_param_ex)

# Plot beta_0, growth intercepts
plots[[1]]

# Plot beta_dbh, growth-dbh slopt
plots[[2]]

# Plot lambdas, competition coefficents
plots[[3]]
}
