% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{posterior_param_ex}
\alias{posterior_param_ex}
\title{Example fit model}
\format{
??
\describe{
\item{??}{Something}
}
}
\usage{
posterior_param_ex
}
\description{
This has posterior parameters for the fit growth model.
}
\examples{
library(dplyr)
library(yardstick)

# Compare model predictions to observation
predictions <-
  predict(posterior_param_ex, newdata = focal_vs_comp_ex) \%>\%
    bind_cols(
      focal_vs_comp_ex \%>\%
        select(focal_ID, focal_sp, dbh, growth) \%>\%
        distinct()
    )
predictions \%>\%
  rmse(truth = growth, estimate = .pred) \%>\%
  pull(.estimate)

# Plot posterior parameters
posterior_param_ex \%>\%
  plot_bayesian_model_parameters()
}
\seealso{
\code{\link{focal_vs_comp_ex}}
}
\keyword{datasets}
