% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict.fe_bayes_lm}
\alias{predict.fe_bayes_lm}
\title{Make predictions based on fitted Bayesian model}
\source{
Closed-form solutions of Bayesian linear regression \url{https://doi.org/10.1371/journal.pone.0229930.s004}
}
\usage{
\method{predict}{fe_bayes_lm}(focal_vs_comp, posterior_param)
}
\arguments{
\item{focal_vs_comp}{data frame from \code{\link{create_focal_vs_comp}}}

\item{posterior_param}{Output of \code{\link{fe_bayes_lm}}: A list of
\verb{\{a_star, b_star, mu_star, V_star\}} posterior hyperparameters}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{growth_hat}
}
\description{
Applies fitted model from \code{\link{fe_bayes_lm}} and
returns posterior predicted values.
}
\examples{
library(dplyr)
library(sf)
library(ggplot2)

# Load in posterior parameter example
# and growth data to compare to
data(posterior_param_ex, ex_growth_df)

predictions <- focal_vs_comp_ex \%>\%
  predict(posterior_param = posterior_param_ex) \%>\%
  right_join(ex_growth_df, by = c("focal_ID" = "ID"))
predictions \%>\%
  ggplot(aes(growth, growth_hat)) +
  geom_point() +
  geom_abline(slope = 1, intercept = 0)
}
\seealso{
\code{\link{fe_bayes_lm}}
}
