% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/re_exported_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{to_any_case}
\alias{rmse}
\alias{sf_polygon}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{sfheaders}{\code{\link[sfheaders]{sf_polygon}}}

  \item{snakecase}{\code{\link[snakecase]{to_any_case}}}

  \item{yardstick}{\code{\link[yardstick]{rmse}}}
}}

