% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_datasets.R
\docType{data}
\name{study_region_bw}
\alias{study_region_bw}
\title{Bigwoods forest study region boundary}
\format{
A \code{sf} spatial features polygon
}
\usage{
study_region_bw
}
\description{
Boundary region for bigwoods defined in terms of (x,y) vertices of a polygon.
}
\examples{
library(ggplot2)
library(sf)

# Convert all 2008 plot stems to sf object
census_2008_bw_sf <- census_2008_bw \%>\%
  st_as_sf(coords = c("gx", "gy"))

# Plot stems with plot boundary
ggplot() +
  geom_sf(data = census_2008_bw_sf, size = 0.25) +
  geom_sf(data = study_region_bw, color = "red", fill = "transparent")
}
\seealso{
\code{\link{census_2008_bw}} and \code{\link{census_2014_bw}}
}
\keyword{datasets}
