% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bigwoods_study_region}
\alias{bigwoods_study_region}
\title{Bigwoods forest study region boundary}
\format{A data frame with 13 rows and 2 variables:
\describe{
\item{x}{x-coordinate (meters from reference point)}
\item{y}{y-coordinate (meters from reference point)}
}}
\usage{
bigwoods_study_region
}
\description{
Boundary region for bigwoods defined in terms of (x,y) vertices of a polygon.
}
\examples{
library(ggplot2)
ggplot(bigwoods, aes(x = x, y = y)) +
  # Mark study region boundary
  geom_path(data = bigwoods_study_region, size = 1) +
  coord_fixed(ratio = 1)
}
\seealso{
\code{\link{families}} \code{\link{define_bigwoods_buffer}}
}
\keyword{datasets}
