% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bw_species}
\alias{bw_species}
\title{Phylogenic groupings and trait based clustering of various tree species}
\format{A data frame with 46 rows and 6 variables:
\describe{
\item{sp}{The code for the species. Link with sp in \code{bw_censusdf1}.}
\item{genus}{Genus}
\item{species}{Species epithet}
\item{latin}{Scientific name}
\item{family}{Family}
\item{trait_group}{Clustering of species based on three traits rather than
their evolutionary relationships. The traits are specific leaf area, maximum
height, and wood density}
}}
\source{
For more information on trait clustering see Allen and Kim 2020. A permutation
test and spatial cross-validation approach to assess models of interspecific competition
between trees. Plos One 15: e0229930. https://doi.org/10.1371/journal.pone.0229930
}
\usage{
bw_species
}
\description{
A date frame mapping the species codes to their common names, scientific
names, and families. This also includes a trait-based clustering of the
species.
}
\seealso{
\code{\link{bw_censusdf2}} \code{\link{bw_censusdf1}}
}
\keyword{datasets}
