% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bw_study_region}
\alias{bw_study_region}
\title{Bigwoods forest study region boundary}
\format{
A \code{sf} spatial features polygon
}
\usage{
bw_study_region
}
\description{
Boundary region for bigwoods defined in terms of (x,y) vertices of a polygon.
}
\examples{
library(ggplot2)
library(sf)

# Convert all 2008 plot stems to sf object
bw_census_2008_sf <- bw_census_2008 \%>\%
  st_as_sf(coords = c("gx", "gy"))

# Plot stems with plot boundary
ggplot() +
  geom_sf(data = bw_census_2008_sf, size = 0.25) +
  geom_sf(data = bw_study_region, color = "red", fill = "transparent")
}
\seealso{
\code{\link{bw_census_2008}}, \code{\link{bw_census_2014}}, \code{\link{define_buffer}}
}
\keyword{datasets}
