% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bw_study_region}
\alias{bw_study_region}
\title{Bigwoods forest study region boundary}
\format{A \code{sf} spatial polygon.
\describe{
\item{x}{x-coordinate (meters from reference point)}
\item{y}{y-coordinate (meters from reference point)}
}}
\usage{
bw_study_region
}
\description{
Boundary region for bigwoods defined in terms of (x,y) vertices of a polygon.
}
\examples{
library(ggplot2)
ggplot(bw_censusdf1, aes(x = gx, y = gy)) +
  # Mark study region boundary
  geom_path(data = bw_study_region, size = 1) +
  coord_fixed(ratio = 1)
}
\seealso{
\code{\link{bw_censusdf1}} \code{\link{define_buffer}}
}
\keyword{datasets}
