% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{comp_bayes_lm_ex}
\alias{comp_bayes_lm_ex}
\title{Example bayesian competition model fit}
\format{
A list subclass containing the following elements:

\describe{
\item{prior_params}{Prior parameters supplied to \code{\link[=comp_bayes_lm]{comp_bayes_lm()}}}
\item{post_params}{Posterior parameters outputted by \code{\link[=comp_bayes_lm]{comp_bayes_lm()}}}
\item{terms}{The formula object used in model fitting}
}
}
\usage{
comp_bayes_lm_ex
}
\description{
This object contains an example fitted bayesian competition
model outputted by \code{\link[=comp_bayes_lm]{comp_bayes_lm()}}.
}
\examples{
library(dplyr)
library(yardstick)

# Compare model predictions to observation
predictions <- focal_vs_comp_ex \%>\%
  mutate(growth_hat = predict(comp_bayes_lm_ex, focal_vs_comp_ex))

predictions \%>\%
  rmse(truth = growth, estimate = growth_hat) \%>\%
  pull(.estimate)

# Plot posterior parameters
comp_bayes_lm_ex \%>\%
  autoplot()
}
\seealso{
\code{\link{focal_vs_comp_ex}}
}
\keyword{datasets}
