% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{compute_growth}
\alias{compute_growth}
\title{Compute growth of trees}
\usage{
compute_growth(census_df1, census_df2, id)
}
\arguments{
\item{census_df1}{A data frame of the first census.}

\item{census_df2}{A data frame of the second (later) census}

\item{id}{Name of variable common to \code{census_df1} and \code{census_df2}
allowing you to join/merge both data frames.}
}
\value{
growth_df An sf data frame with \code{growth}: average annual growth in dbh.
}
\description{
Based on two tree censuses compute the average annual growth in dbh for all
trees that were alive at (TODO: fill this in).
}
\examples{
library(dplyr)
library(stringr)
# Load in data from two forst censuses
data(census_df1_ex, census_df2_ex)
# Filter out resprouts in second census
census_df2_ex_no_r <- census_df2_ex \%>\%
 filter(!str_detect(codes, 'R'))
id <- 'ID'
ex_growth_df <- compute_growth(census_df1_ex, census_df2_ex_no_r, id)
}
