% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{create_bayesian_model_data}
\alias{create_bayesian_model_data}
\title{Create input data frame for Bayesian regression}
\usage{
create_bayesian_model_data(focal_vs_comp, run_shuffle = FALSE)
}
\arguments{
\item{focal_vs_comp}{data frame from \code{\link{create_focal_vs_comp}}}

\item{run_shuffle}{boolean as to whether to run permutation test shuffle of
competitor tree species within a particular focal_ID}
}
\value{
Data frame that can be used for lm()
}
\description{
This function is used both by \code{\link{comp_bayes_lm}} and \code{\link{predict.comp_bayes_lm}}
}
\examples{
1 + 1
}
