% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{create_focal_vs_comp}
\alias{create_focal_vs_comp}
\title{Create focal versus competitor trees data frame}
\usage{
create_focal_vs_comp(growth_df, max_dist, cv_grid_sf, id)
}
\arguments{
\item{growth_df}{A \code{\link{compute_growth}} output converted to \code{sf} object}

\item{max_dist}{Distance to determine which neighboring trees to a focal tree are competitors.}

\item{cv_grid_sf}{An sf object of a \code{blockCV} block output}

\item{id}{A character string of the variable name in \code{growth_df} uniquely identifying each tree}
}
\value{
\code{focal_vs_comp} data frame of all focal trees and for each focal
tree all possible competitor trees. In particular, for each competitor tree
we compute the \href{https://en.wikipedia.org/wiki/Basal_area}{basal area}
(in meters-squared) based on the \code{dbh1} variable from the first census
(assumed to be in cm).
}
\description{
"Focal versus competitor trees" data frames are the main data
frame used for analysis. "Focal trees" are all trees that satisfy the
following criteria
\tabular{l}{
\enumerate{
\item Were alive at both censuses \cr
\item Were not part of the study region's buffer as computed by \code{\link{add_buffer_variable}} \cr
\item Were not a resprout at the second census. Such trees should be coded as
\code{"R"} in the \code{codes2} variable (OK if a resprout at first census)
}
For each focal tree, "competitor trees" are all trees that (1) were alive at
the first census and (2) within \code{max_dist} distance of the focal tree.
}
}
\note{
In order to speed computation, in particular of distances between all
focal/competitor tree pairs, we use the cross-validation \code{blockCV}
object to divide the study region into smaller subsets.
}
\examples{
library(ggplot2)
library(dplyr)
library(stringr)
library(sf)
library(sfheaders)
library(tibble)

# Create fold information sf object. TODO: clean this
cv_grid_ex <-
  tibble(
    # Study region boundary
    x = c(0, 0, 5, 5),
    y = c(0, 5, 5, 0)
  ) \%>\%
  # Convert to sf object
  sf_polygon() \%>\%
  mutate(folds = "1")

# Plot example data. Observe for max_dist = 1.5, there are 6 focal vs comp pairs:
# 1. focal 1 vs comp 2
# 2. focal 2 vs comp 1
# 3. focal 2 vs comp 3
# 4. focal 3 vs comp 2
# 5. focal 4 vs comp 5
# 6. focal 5 vs comp 4
ggplot() +
  geom_sf(data = cv_grid_ex, fill = "transparent") +
  geom_sf_label(data = growth_df_ex, aes(label = ID))

# Return corresponding data frame
growth_df_ex \%>\%
  create_focal_vs_comp(max_dist = 1.5, cv_grid_sf = cv_grid_ex, id = "ID")

# Load in growth_df with spatial data
# See ?ex_growth_df for attaching spatial data to growth_df
data(ex_growth_df_spatial)
# Load in cv_grid
data(ex_cv_grid_sf)

focal_vs_comp_ex <- ex_growth_df_spatial \%>\%
  create_focal_vs_comp(max_dist = 1, cv_grid_sf = ex_cv_grid_sf, id = "ID")
}
\seealso{
\code{\link{focal_vs_comp_distance}}
}
