% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{create_focal_vs_comp}
\alias{create_focal_vs_comp}
\title{Create main data frame for analysis}
\usage{
create_focal_vs_comp(
  growth_df,
  max_dist,
  model_specs,
  cv_grid,
  id,
  plot_folds = FALSE
)
}
\arguments{
\item{growth_df}{\code{\link{compute_growth}} data frame}

\item{max_dist}{Maximum distance from a focal tree for another tree to be
considered a competitor tree}

\item{model_specs}{from \code{\link{get_model_specs}}}
}
\value{
\code{focal_vs_comp} data frame
}
\description{
Create main data frame for analysis
}
\examples{
1+1
}
\seealso{
\code{\link{define_cv_grid}} and \code{\link{get_model_specs}}
}
