% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_bayesian_model}
\alias{fit_bayesian_model}
\title{Fit bayesian model}
\usage{
fit_bayesian_model(focal_vs_comp, run_shuffle = FALSE)
}
\arguments{
\item{focal_vs_comp}{from \code{\link{create_focal_vs_comp}}}

\item{run_shuffle}{boolean}

\item{focal_trees}{training set}
}
\value{
Posterior parameter values
}
\description{
Fit bayesian model
}
\examples{
1+1
}
