% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_bayesian_model}
\alias{fit_bayesian_model}
\title{Fit Bayesian competition model}
\source{
Closed-form solutions of Bayesian linear regression \url{https://doi.org/10.1371/journal.pone.0229930.s004}
}
\usage{
fit_bayesian_model(focal_vs_comp, prior_param = NULL, run_shuffle = FALSE)
}
\arguments{
\item{focal_vs_comp}{data frame from \code{\link{create_focal_vs_comp}}}

\item{prior_param}{A list of \verb{\{a_0, b_0, mu_0, V_0\}} prior hyperparameters}

\item{run_shuffle}{boolean as to whether to run permutation test shuffle of
competitor tree species within a particular focal_ID}
}
\value{
A list of \verb{\{a_star, b_star, mu_star, V_star\}} posterior hyperparameters
}
\description{
Fit a Bayesian linear regression model with interactions terms where \deqn{y = X \beta + \epsilon}
\tabular{ll}{
\eqn{\mu} \tab mean hyperparameter vector for \eqn{\beta} of length \eqn{p + 1} \cr
\eqn{V} \tab covariance hyperparameter matrix for \eqn{\beta} of dimension \eqn{(p + 1) x (p + 1)} \cr
\eqn{a} \tab shape hyperparameter for \eqn{\sigma^2 > 0} \cr
\eqn{b} \tab scale hyperparameter for \eqn{\sigma^2 > 0}\cr
}
}
\examples{
1+1
}
\seealso{
\code{\link{predict_bayesian_model}}
}
