% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_cv_fold_info}
\alias{get_cv_fold_info}
\title{Return information on each fold in crossvalidation grid}
\usage{
get_cv_fold_info(forest, cv_fold_size)
}
\arguments{
\item{forest}{data frame containing tree data}

\item{cv_fold_size}{grid size for spatial crossvalidation fold}
}
\value{
A "nested" dataframe with the fold number, its rough centroid, and
all its neighbors in a nested list.
}
\description{
Return information on each fold in crossvalidation grid
}
\examples{
library(ggplot2)
library(dplyr)
\dontrun{

# Define crossvalidation folds/grid
bigwoods <- bigwoods \%>\%
  define_cv_grid(cv_fold_size = 100) \%>\%
  mutate(fold = factor(fold))

# CV fold/grid info
folds <- bigwoods \%>\%
  get_cv_fold_info(cv_fold_size = 100)

# Plot fold numbers
ggplot(data = bigwoods, aes(x = x, y = y)) +
  geom_point(aes(col = fold)) +
  coord_fixed(ratio = 1) +
  geom_text(data = folds, aes(x = x, y = y, label = fold), size = 10)

# Check that neighbors are correct:
library(stringr)
library(tidyr)
neighbors <- folds \%>\%
  unnest() \%>\%
  group_by(fold) \%>\%
  summarize(neighbors = str_c(neighbor, collapse = ","))
neighbors
}
}
