% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{growth_toy}
\alias{growth_toy}
\title{Example input data for \code{\link[=create_focal_vs_comp]{create_focal_vs_comp()}}}
\format{
A \code{sf} spatial features polygon
}
\usage{
growth_toy
}
\description{
An example \code{sf} of type generated by \code{\link[=compute_growth]{compute_growth()}}
}
\seealso{
Other example data objects: 
\code{\link{blocks_ex}},
\code{\link{census_1_ex}},
\code{\link{census_2008_bw}},
\code{\link{census_2014_bw}},
\code{\link{census_2_ex}},
\code{\link{comp_bayes_lm_ex}},
\code{\link{focal_vs_comp_ex}},
\code{\link{growth_ex}},
\code{\link{growth_spatial_ex}},
\code{\link{species_bw}},
\code{\link{study_region_bw}},
\code{\link{study_region_ex}}
}
\concept{example data objects}
\keyword{datasets}
