% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{plot_posterior_parameters}
\alias{plot_posterior_parameters}
\title{Plot beta_0 parameters}
\usage{
plot_posterior_parameters(posterior_param, species_list)
}
\arguments{
\item{posterior_param}{Output of \code{\link{fit_bayesian_model}}}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{growth_hat}
}
\description{
Plot beta_0 parameters
}
\examples{
1+1
}
