% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict.fe_bayes_lr}
\alias{predict.fe_bayes_lr}
\title{Make predictions based on fitted Bayesian model}
\source{
Closed-form solutions of Bayesian linear regression \url{https://doi.org/10.1371/journal.pone.0229930.s004}
}
\usage{
\method{predict}{fe_bayes_lr}(object, ...)
}
\arguments{
\item{object}{data frame from \code{\link{create_focal_vs_comp}}}

\item{...}{Currently ignored, only included for consistency with generic.}
}
\value{
A tibble with column \code{.pred}
}
\description{
Applies fitted model from \link{fit.fe_data} and
returns posterior predicted values.
}
\examples{
library(dplyr)
library(sf)
library(ggplot2)

# Load in posterior parameter example
# and growth data to compare to
data(posterior_param_ex, ex_growth_df)

predictions <- focal_vs_comp_ex \%>\%
  predict(posterior_param = posterior_param_ex) \%>\%
  right_join(ex_growth_df, by = c("focal_ID" = "ID"))
predictions \%>\%
  ggplot(aes(growth, .pred)) +
  geom_point() +
  geom_abline(slope = 1, intercept = 0)
}
\seealso{
\link{fit.fe_data}
}
