% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{predict_bayesian_model}
\alias{predict_bayesian_model}
\title{Make predictions based on bayesian model}
\usage{
predict_bayesian_model(focal_vs_comp, model_formula, posterior_param)
}
\arguments{
\item{focal_vs_comp}{from \code{\link{create_focal_vs_comp}}}

\item{posterior_param}{Output of \code{\link{fit_bayesian_model}}}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{growth_hat}
}
\description{
Make predictions based on bayesian model
}
\examples{
1+1
}
