% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{run_cv}
\alias{run_cv}
\title{Run the bayesian model with spatial cross validation}
\usage{
run_cv(
  focal_vs_comp,
  model_formula,
  max_dist,
  cv_grid,
  run_shuffle = FALSE,
  prior_hyperparameters = NULL,
  all_folds = TRUE
)
}
\arguments{
\item{focal_vs_comp}{from \code{\link{create_focal_vs_comp}}}

\item{max_dist}{distance of competitive neighborhood}

\item{cv_grid}{length of the cross validation grid}

\item{run_shuffle}{boolean as to whether to run permutation test shuffle}

\item{prior_hyperparameters}{list of a0, b0, mu_0 and V_0}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{growth_hat}
}
\description{
Run the bayesian model with spatial cross validation
}
\examples{
1+1
}
