% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{run_cv}
\alias{run_cv}
\title{Run the bayesian model with spatial cross validation}
\usage{
run_cv(
  focal_vs_comp,
  comp_dist,
  cv_grid,
  prior_param = NULL,
  run_shuffle = FALSE
)
}
\arguments{
\item{focal_vs_comp}{data frame from \code{\link{create_focal_vs_comp}}}

\item{comp_dist}{Distance to determine which neighboring trees to a focal tree are competitors.}

\item{cv_grid}{\code{sf} polygon output from \code{\link[blockCV]{spatialBlock}}}

\item{prior_param}{A list of \verb{\{a_0, b_0, mu_0, V_0\}} prior hyperparameters}

\item{run_shuffle}{boolean as to whether to run permutation test shuffle of
competitor tree species within a particular focal_ID}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{growth_hat}
}
\description{
Run cross-validation
}
\examples{
1 + 1
}
