% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{run_cv}
\alias{run_cv}
\title{Run the bayesian model with spatial cross validation}
\usage{
run_cv(
  focal_vs_comp,
  model_formula,
  max_dist,
  cv_grid,
  run_shuffle = FALSE,
  prior_hyperparameters = NULL,
  all_folds = TRUE
)
}
\arguments{
\item{focal_vs_comp}{data frame from \code{\link{create_focal_vs_comp}}}

\item{model_formula}{model formula of linear regression}

\item{max_dist}{distance of competitive neighborhood}

\item{cv_grid}{length of the cross validation grid}

\item{run_shuffle}{boolean as to whether to run permutation test shuffle of
competitor tree species within a particular focal_ID}

\item{prior_hyperparameters}{A list of \verb{\{a_0, b_0, mu_0, V_0\}} prior hyperparameters}
}
\value{
\code{focal_vs_comp} with new column of predicted \code{growth_hat}
}
\description{
Run the bayesian model with spatial cross validation
}
\examples{
1+1
}
